rename: ResponseCodeDescription 리패키징
